Serve assets inline when not in browser cache
Each inline asset has a corresponding custom element that will cache its
contents using the Cache API and add it to an asset catalog stored in
IndexedDB. The service worker will then intercept navigation requests
and add an `X-Asset-Digest` header containing a list of cached bundles
with the URL of their cached versions as read from the asset catalog.

The server will inline any asset bundles that aren’t present in the
header (if the header itself is present at all), or which have an
outdated version in the header. Asset bundles present in the header at
the current version will instead be served using an external reference
so that they can be read from the browser cache.